% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_7_day.R
\name{get_7_day}
\alias{get_7_day}
\title{Gets eligible patients based on 7 day visit criteria}
\usage{
get_7_day(df)
}
\arguments{
\item{df}{the name of the dataset produces from "process_data()" function}
}
\value{
Dataframe containing only patients with missed visits from exactly 7 days ago
}
\description{
Processes data further and reduces dataset to patients eligibility based on their appoitnment having occured 7 days prior.
}
\examples{
get_7_day(df)
}
