% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_df.R
\name{process_df}
\alias{process_df}
\title{Process dataframe}
\usage{
process_df(df)
}
\arguments{
\item{df}{the name of the dataframe stored from "import_df()"}
}
\value{
cleaned/process version of the dataframe
}
\description{
performs basic precursory cleaning of the dataframe from "import_df()". Removes duplicates, labels sites, adds redcap variables and re-orders variables
}
\examples{
process_df(df)
}
